Migrated to PodMutator from DeploymentMutator to avoid blocking k8s deployments before KFServing is installed
