+++ pkg/webhook/admission/pod/accelerator_injector_test.go
@@ -13,14 +13,13 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package deployment
+package pod
 
 import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/kubeflow/kfserving/pkg/constants"
-	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -28,83 +27,67 @@ import (
 
 func TestAcceleratorInjector(t *testing.T) {
 	scenarios := map[string]struct {
-		original *appsv1.Deployment
-		expected *appsv1.Deployment
+		original *v1.Pod
+		expected *v1.Pod
 	}{
 		"AddGPUSelector": {
-			original: &appsv1.Deployment{
+			original: &v1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Name: "deployment",
 					Annotations: map[string]string{
 						constants.KFServiceGKEAcceleratorAnnotationKey: "nvidia-tesla-v100",
 					},
 				},
-				Spec: appsv1.DeploymentSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Resources: v1.ResourceRequirements{
-									Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
-								},
-							}},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Resources: v1.ResourceRequirements{
+							Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
 						},
-					},
+					}},
 				},
 			},
-			expected: &appsv1.Deployment{
+			expected: &v1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Name: "deployment",
 					Annotations: map[string]string{
 						constants.KFServiceGKEAcceleratorAnnotationKey: "nvidia-tesla-v100",
 					},
 				},
-				Spec: appsv1.DeploymentSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							NodeSelector: map[string]string{
-								GkeAcceleratorNodeSelector: "nvidia-tesla-v100",
-							},
-							Containers: []v1.Container{{
-								Resources: v1.ResourceRequirements{
-									Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
-								},
-							}},
-						},
+				Spec: v1.PodSpec{
+					NodeSelector: map[string]string{
+						GkeAcceleratorNodeSelector: "nvidia-tesla-v100",
 					},
+					Containers: []v1.Container{{
+						Resources: v1.ResourceRequirements{
+							Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
+						},
+					}},
 				},
 			},
 		},
 		"DoNotAddGPUSelector": {
-			original: &appsv1.Deployment{
+			original: &v1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Name: "deployment",
 				},
-				Spec: appsv1.DeploymentSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Resources: v1.ResourceRequirements{
-									Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
-								},
-							}},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Resources: v1.ResourceRequirements{
+							Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
 						},
-					},
+					}},
 				},
 			},
-			expected: &appsv1.Deployment{
+			expected: &v1.Pod{
 				ObjectMeta: metav1.ObjectMeta{
 					Name: "deployment",
 				},
-				Spec: appsv1.DeploymentSpec{
-					Template: v1.PodTemplateSpec{
-						Spec: v1.PodSpec{
-							Containers: []v1.Container{{
-								Resources: v1.ResourceRequirements{
-									Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
-								},
-							}},
+				Spec: v1.PodSpec{
+					Containers: []v1.Container{{
+						Resources: v1.ResourceRequirements{
+							Limits: v1.ResourceList{constants.NvidiaGPUResourceType: resource.MustParse("1")},
 						},
-					},
+					}},
 				},
 			},
 		},
@@ -114,17 +97,16 @@ func TestAcceleratorInjector(t *testing.T) {
 		InjectGKEAcceleratorSelector(scenario.original)
 		// cmd.Diff complains on ResourceList when Nvidia is key. Objects are explicitly compared
 		if diff := cmp.Diff(
-			scenario.expected.Spec.Template.Spec.NodeSelector,
-			scenario.original.Spec.Template.Spec.NodeSelector,
+			scenario.expected.Spec.NodeSelector,
+			scenario.original.Spec.NodeSelector,
 		); diff != "" {
 			t.Errorf("Test %q unexpected result (-want +got): %v", name, diff)
 		}
 		if diff := cmp.Diff(
-			scenario.expected.Spec.Template.Spec.Tolerations,
-			scenario.original.Spec.Template.Spec.Tolerations,
+			scenario.expected.Spec.Tolerations,
+			scenario.original.Spec.Tolerations,
 		); diff != "" {
 			t.Errorf("Test %q unexpected result (-want +got): %v", name, diff)
 		}
-
 	}
 }
